Fast Track: Parse and compare in-vm metadata, status upload blob and artifacts profile (#2411)
* Fast Track: Parse and compare in-vm metadata, status upload blob and artifacts profile

* Created test_extensions_goal_state_from_vm_settings.py

* Fix pylint warning

* Fix pylint warning

* Review feedback

* typo
