+++ tests/data/wire/ext_conf_invalid_vm_metadata.xml
@@ -25,5 +25,5 @@
       </Plugin>
    </PluginSettings>
    <StatusUploadBlob statusBlobType="BlockBlob">https://test.blob.core.windows.net/vhds/test-cs12.test-cs12.test-cs12.status?sr=b&amp;sp=rw&amp;se=9999-01-01&amp;sk=key1&amp;sv=2014-02-14&amp;sig=hfRh7gzUE7sUtYwke78IOlZOrTRCYvkec4hGZ9zZzXo</StatusUploadBlob>
-   <InVMGoalStateMetaData inSvdSeqNo="2" createdOnTicks="0" activityId="555e551c-600e-4fb4-90ba-8ab8ec28eccc" correlationId="400de90b-522e-491f-9d89-ec944661f531" />
+   <InVMGoalStateMetaData inSvdSeqNo="2" createdOnTicks="0" activityId="" correlationId="" />
 </Extensions>
