+++ tests/protocol/test_hostplugin.py
@@ -34,7 +34,7 @@ from azurelinuxagent.common.utils import restutil
 from azurelinuxagent.common.version import AGENT_VERSION, AGENT_NAME
 from tests.protocol.mocks import mock_wire_protocol, HttpRequestPredicates, MockHttpResponse
 from tests.protocol.mockwiredata import DATA_FILE, DATA_FILE_NO_EXT
-from tests.tools import AgentTestCase, PY_VERSION_MAJOR, Mock, patch
+from tests.tools import AgentTestCase, PY_VERSION_MAJOR, Mock, PropertyMock, patch
 
 
 hostplugin_status_url = "http://168.63.129.16:32526/status"
@@ -152,9 +152,9 @@ class TestHostPlugin(HttpRequestPredicates, AgentTestCase):
         with mock_wire_protocol(DATA_FILE_NO_EXT) as protocol:
             # These tests use mock wire data that don't have any extensions (extension config will be empty).
             # Populate the upload blob and set an initial empty status before returning the protocol.
-            extensions_goal_state = protocol.client.get_extensions_goal_state()
-            extensions_goal_state._status_upload_blob = sas_url
-            extensions_goal_state._status_upload_blob_type = page_blob_type
+            protocol.client._extensions_goal_state = Mock(wraps=protocol.client._extensions_goal_state)
+            type(protocol.client._extensions_goal_state).status_upload_blob = PropertyMock(return_value=sas_url)
+            type(protocol.client._extensions_goal_state).status_upload_blob_type = PropertyMock(return_value=page_blob_type)
 
             status = restapi.VMStatus(status="Ready", message="Guest Agent is running")
             protocol.client.status_blob.set_vm_status(status)
